changed leetcode easy challenges

+++ Leetcode/easy/2278. Percentage of Letter in String.py
@@ -1,12 +1,9 @@
-# Time to write all of below including tests, explanation and time and aux
-# and total space: 1 min
 
-# Problem: https://leetcode.com/problems/percentage-of-letter-in-string/description/
 
 class Solution:
     def percentageLetter(self, s: str, letter: str) -> int:
-        # Time: O(n), n = len(s)
+        # Time: O(n)
         # Space: O(1)
-        return int((100 * s.count(letter)) // len(s))
+        return (100 * s.count(letter)) // len(s)
